Safari-609.1.11.1/LayoutTests/crypto/subtle/ecdsa-import-key-verify-p521-expected.txt

Test ECDSA verifying operation with a P-521 public key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS verified is true
PASS successfullyParsed is true

TEST COMPLETE
